move polyfills to where their methods are used
