+++ src/core/util/functional.ts
@@ -25,14 +25,32 @@ export function range(startOrSize: number, end?: number) {
   return end ? [...Array(end).keys()].slice(startOrSize) : [...Array(startOrSize).keys()]
 }
 
+export function unique<T>(items: T[] | undefined): T[] {
+  return Array.from(new Set(items))
+}
+
+// remove this after node 22 is lts
+if (!Set.prototype.intersection) {
+  Set.prototype.intersection = function (other) {
+    const r = new Set()
+    for (const v of this) if (other.has(v)) r.add(v)
+    return r
+  }
+}
+
 export function intersection<T>(first: T[] | undefined, second: T[]): T[] {
   const set1 = new Set(first)
   const set2 = new Set(second)
   return Array.from(set1.intersection(set2))
 }
 
-export function unique<T>(items: T[] | undefined): T[] {
-  return Array.from(new Set(items))
+// remove this after node 22 is lts
+if (!Set.prototype.difference) {
+  Set.prototype.difference = function (other) {
+    const r = new Set()
+    for (const v of this) if (!other.has(v)) r.add(v)
+    return r
+  }
 }
 
 export function difference<T>(first: T[] | undefined, second: T[]): T[] {
